+++ src/main/java/fm/unit/data/ObjectFactory.java
@@ -6,7 +6,7 @@
 //
 
 
-package unitfm.data;
+package fm.unit.data;
 
 import javax.xml.bind.JAXBElement;
 import javax.xml.bind.annotation.XmlElementDecl;
@@ -17,7 +17,7 @@ import javax.xml.namespace.QName;
 /**
  * This object contains factory methods for each 
  * Java content interface and Java element interface 
- * generated in the unitfm.data package. 
+ * generated in the fm.unit.data package.
  * <p>An ObjectFactory allows you to programatically 
  * construct new instances of the Java representation 
  * for XML content. The Java representation of XML 
@@ -36,7 +36,7 @@ public class ObjectFactory {
     private final static QName _SystemOut_QNAME = new QName("", "system-out");
 
     /**
-     * Create a new ObjectFactory that can be used to create new instances of schema derived classes for package: unitfm.data
+     * Create a new ObjectFactory that can be used to create new instances of schema derived classes for package: fm.unit.data
      * 
      */
     public ObjectFactory() {
