+++ src/main/java/fm/unit/data/Properties.java
@@ -6,7 +6,7 @@
 //
 
 
-package unitfm.data;
+package fm.unit.data;
 
 import java.util.ArrayList;
 import java.util.List;
