+++ src/main/resources/db/migration/V0__Boostraph.sql
@@ -1,14 +1,19 @@
+CREATE TABLE organizations (organization_id SERIAL PRIMARY KEY, name text NOT NULL);
+CREATE TABLE repositories (repository_id SERIAL PRIMARY KEY, name text NOT NULL);
+
 /* A report of a build */
 CREATE TABLE reports (
-    id SERIAL PRIMARY KEY,
-    commit text NOT NULL,
-    key text NOT NULL /*S3 like key.*/
+    report_id SERIAL PRIMARY KEY,
+    organization_id INTEGER REFERENCES organizations(organization_id),
+    repository_id INTEGER REFERENCES organizations(organization_id),
+    commit_hash text NOT NULL,
+    prefix text NOT NULL
 );
 
 /* Testsuite files. Basically all JUnit files of a build. */
 CREATE TABLE testsuites (
-    id SERIAL PRIMARY KEY,
-    report_id integer REFERENCES reports,
-    name text,
+    testsuite_id SERIAL PRIMARY KEY,
+    report_id INTEGER REFERENCES reports(report_id),
+    filename text NOT NULL,
     payload xml NOT NULL
-);
+)
